requirements.txt
aiohttp==3.12.15
dateparser==1.2.2
python-dotenv==1.1.1
fastapi==0.118.0
munch==4.0.0
PyYAML==6.0.3
streamlit==1.50.0
substack-api==1.1.1
twikit==2.3.3
uvicorn==0.37.0
youtube-transcript-api==1.2.2
